feat: Add NetworkList getter and redo NetworkList tests (#3690)
## Purpose of this PR

Add setter onto `NetworkList` that allows an update to be "forced" even
if the value was not changed.

fixes: #3657 

### Jira ticket
MTT-13366

### Changelog

- Added: A `Set` function onto the `NetworkList` that takes an optional
parameter that will force the update to be processed and will ignore the
equality check.

## Documentation

- No documentation changes or additions were necessary.

## Testing & QA (How your changes can be verified during release
Playtest)

### Functional Testing

- [ ] `Manual testing done`

_Automated tests:_
- [x] `Covered by existing automated tests`
- [x] `Covered by new automated tests`

_Does the change require QA team to:_

- [ ] `Review automated tests`?
- [ ] `Execute manual tests`?
- [ ] `Provide feedback about the PR`?

If any boxes above are checked the QA team will be automatically added
as a PR reviewer.

## Backports

This is a 2.x only change

---------

Co-authored-by: Noel Stephens <noel.stephens@unity3d.com>