fix: disconnect reason should not duplicate any information (#3757)
* fix

Fixing issue with duplicated disconnect reasons for all cases.

* test - update

Making adjustments to the validation logic checking to determine if a server's disconnect reason is included in the NetworkManager.DisconnectReaon property.

* style

Removing a trailing whitespace

* test - update

One more area that needs to check for the server reason but not be a 1:1 due to the additional disconnect information before it.