fix: disabled NetworkBehaviours vs disabled GameObjects and tests excluded by lack of MP tools (#3761)
* fix

Use the GameObject.activeInHierarchy fix.
Allow disabled NetworkBehaviour with a disabled GameObject to internally start pre-spawn but exit before invoking protected method.

* test - update

Re-factoring this test to align with the previously committed fix.

* tests - enable

Now that we have included the Mutliplayer Tools package in the test project, we needed to remove the !MULTIPLAYER_TOOLS check to assure these tests are actually being run.