test fix
Fixing potential cause for instabilities on this set of tests due to the initial network prefab transform settings not being applied to the prefab but applied to the spawned instances which could be causing issues since they were being applied to the non-authority instances.

This fix moves the applied test NetworkTrransform settings to the player prefab prior to spawning any instances.