fix: 2.7.0 test instabilities (#3768)
* test - fix

Resolving the PeerDisconnectCallbackTests instability where sometimes the disconnect notification is processed prior to the targeted client's NetworkManager having completely disconnected and shutdown (this would only have a chance of happening if server disconnects the client).

Hoping to resolve the instability with the ParentingInSceneObjectsTests resulting in a false failure.

* test fix

Fixing potential cause for instabilities on this set of tests due to the initial network prefab transform settings not being applied to the prefab but applied to the spawned instances which could be causing issues since they were being applied to the non-authority instances.

This fix moves the applied test NetworkTrransform settings to the player prefab prior to spawning any instances.