update
Converting all legacy ClientRpc and ServerRpc usages over to the newer universal RPC format.
Using a pre-allocated Persistent RpcTargetGroup as opposed to the List.
When using `NetworkAnimator.SetTrigger` with a client-server network topology and it is the server or host invoking the method, then go ahead and set the trigger since it no longer sends itself an RPC/no longer invokes the InternalSetTrigger via RPC.