fix: RPC metrics name table is empty (#3782)
* fix

Improving our defines that determine if metrics should be gathered.
1st pass attempt at unifying this process.
The primary change is that if the __rpc_name_table is not built (yet) when an RPC is invoked that it will attempt to build it for that specific NetworkBehaviour prior to attempting to log RPC metrics.

* fix

Missed one set of defines check.

* fix

Removing reference to NetworkBehaviour.__rpc_name_table as it is not used any longer within the codegen process.

* fix

Some additional areas in codegen that requires the extended define wrap.
Wrapping the RPC invocation tests (we will need to find a better way to get the name for this test).

* refactor

Making sure we still have the NetworkManager.__rpc_name_table as that has become part of the public API.

* fix

This fixes the issue with invoking RPCs on other child NetworkBehaviours during OnNetworkSpawn.

* test

Adding a test to validate this specific issue is fixed.

* style

Removing trailing whitespace.

* update

Marking newly added methods for aggressive in-lining.

* update

Adding change log entry.

* style

Fixing a comment.
Adding additional comments.

* style

Making sure the correct ValidateRpcMessageMetrics name is used when logging an error.

* style

Updated and added some comments to the region of code that contains the primary fix.

* fix

Adding target client that is being sent to.

* update

Adding AggressiveInlining to updated method.

---------

Co-authored-by: Unity Netcode CI <74025435+netcode-ci-service@users.noreply.github.com>
Co-authored-by: Noellie Vélez <noellie.velez@unity3d.com>