refactor and fix
Fixing several issues with the disconnect flow.
Not removing the transport id until later. The server now only does this within OnClientDisconnectFromServer.
When disconnecting a client with a reason, the client is not completely disconnected until the end of the frame when the server has finished sending all pending messages.
Fixing some missed MessageDeliveryType replacements.
Removing the MockSkippingApproval as it is no longer needed with the fixes in place.