refactor - EntityId
Since MP Tools subscribes to the TransportInitialized and TransportDisposed methods and we have to vet against packages with dependencies, I am putting a temporary work-around to get past this issue so we can land the EntityId fix.
Once MP Tools is updated for EntityId, we can re-enable the original fix.