fix: client and server disconnect warnings (#3786)
* fix

This fix prevents the warning message from being logged on the client side when shutting down from the client side and the developer log level is set.
This fix also assures that a server or host does not send itself a disconnect message.
This fix separated the transport id clean up between client and host/server upon handling a disconnect event which assures the server/host will have access to this transport id when processing a client that disconnected.

* update

Adding change log entries.

* test

Adding test to check that no warning messages are logged upon a client disconnecting itself from a session.

* refactor and fix

Fixing several issues with the disconnect flow.
Not removing the transport id until later. The server now only does this within OnClientDisconnectFromServer.
When disconnecting a client with a reason, the client is not completely disconnected until the end of the frame when the server has finished sending all pending messages.
Fixing some missed MessageDeliveryType replacements.
Removing the MockSkippingApproval as it is no longer needed with the fixes in place.

* test

Did a bunch of cleanup on several connection tests to improve stability and use more recent integration test features.
ClientOnlyConnectionTests reduced the transport connection timeout and increased the timeout helper.
ConnectionApprovalTimeoutTests needed to have the server and client timeout values vary depending upon the test type.
PeerDisconnectCallbackTests needed to trap for the peer disconnecting.

* style

Removing whitespaces

* update

adding one more change log entry